remove separate file for providers and move content to the module file
