+++ projects/core/src/occ/config-loader/occ-config-loader.module.ts
@@ -1,5 +1,32 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
-import { providers } from './occ-config-loader.providers';
+import {
+  ConfigInitializer,
+  CONFIG_INITIALIZER,
+} from '../../config/config-initializer/config-initializer';
+import { SiteContextConfig } from '../../site-context/config/site-context-config';
+import { BASE_SITE_CONTEXT_ID } from '../../site-context/providers/context-ids';
+import { OccConfigLoaderService } from './occ-config-loader.service';
+
+/**
+ * Initializes the Spartacus config asynchronously basing on the external config
+ *
+ * @internal
+ */
+export function initConfig(
+  configLoader: OccConfigLoaderService,
+  config: SiteContextConfig
+): ConfigInitializer {
+  /**
+   * Load config for `context` from backend only when there is no static config for `context.baseSite`
+   */
+  if (!config.context || !config.context[BASE_SITE_CONTEXT_ID]) {
+    return {
+      scopes: ['context', 'i18n.fallbackLang'],
+      configFactory: () => configLoader.loadConfig(),
+    };
+  }
+  return null;
+}
 
 /**
  * Re-provides the external config chunk given before Angular bootstrap
@@ -9,7 +36,14 @@ export class OccConfigLoaderModule {
   static forRoot(): ModuleWithProviders<OccConfigLoaderModule> {
     return {
       ngModule: OccConfigLoaderModule,
-      providers,
+      providers: [
+        {
+          provide: CONFIG_INITIALIZER,
+          useFactory: initConfig,
+          deps: [OccConfigLoaderService, SiteContextConfig],
+          multi: true,
+        },
+      ],
     };
   }
 }
